保证经纬度的精度，float改为double

+++ BaiduMap/Response/Models/Location.cs
@@ -3,7 +3,7 @@ namespace BaiduMap.Response.Models
 {
     public class Location
     {
-        public float? Lat { get; set; }
-        public float? Lng { get; set; }
+        public double? Lat { get; set; }
+        public double? Lng { get; set; }
     }
 }
